added linear_trajectory, added necessary models, changed module name to trajectories_api, added readme
