+++ blender3d_intergration/blender_python_working/README.md
@@ -1,0 +1,17 @@
+# What do files here do
+
+## execute_script.py
+
+This is a script that You should import to Blender3d
+and execute. This script will take all trajectories 
+one by one, select some object and execute commands from
+trajectory, which will mostly just place object in space,
+render and save image.
+
+Trajectories/objects/paths are all defined in file.
+
+## blender_commands_generator.py
+
+This script generates blender commands which mostly consist
+of commands to set the scene and commands to place objects
+in space according to trajectory.
