+++ blender3d_intergration/interfaces.py
@@ -1,49 +1,58 @@
 from abc import abstractmethod, ABC
 
-from blender3d_intergration.object_trajectory_calculation.models import Trajectory, \
-    Coordinates, CircularTrajectoryMotionParameters, CircularTrajectoryMotionInitialParameters, FramesPerSecond
+from matplotlib import pyplot as plt
+
+from blender3d_intergration.trajectories_api.models import Trajectory, \
+    Coordinates, CircularTrajectoryMotionParameters, CircularTrajectoryMotionInitialParameters, FramesPerSecond, \
+    LinearTrajectoryMotionInitialParameters, LinearTrajectoryMotionParameters
 
 
-class BasicBlenderTrajectoryCalculator:
-    trajectory = []
-
-    def get_trajectory(self):
-        return self.trajectory
-
-    @abstractmethod
-    def set_path_parameters(self):
-        raise NotImplementedError
-
-    @abstractmethod
-    def set_start_coordinates(self, start_x: int, start_y: int, start_z: int):
-        raise NotImplementedError
-
-    @abstractmethod
-    def set_fps(self, fps: int):
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_trajectory_as_blender_script(self):
-        raise NotImplementedError
-
-    @abstractmethod
-    def calculate_trajectory(self) -> None:
-        raise NotImplementedError
-
-    @abstractmethod
-    def __calculate_coordinates(self, frame: int) -> Coordinates:
-        raise NotImplementedError
-
-    @abstractmethod
-    def plot(self):
-        raise NotImplementedError
-
-
-class CircleTrajectory(BasicBlenderTrajectoryCalculator, ABC):
+class TrajectoryCalculator:
     _trajectory: Trajectory
+    _fps: FramesPerSecond
+
+    def calculate_trajectory(self) -> None:
+        self._trajectory.clear_coordinates()
+
+        end_frame = self._calculate_end_frame()
+
+        for frame in range(0, end_frame + 1):
+            coordinates_for_frame = self._calculate_coordinates(frame)
+            self._trajectory.add_coordinates(coordinates_for_frame)
+
+    def plot(self):
+        x = self._trajectory.get_x_positions()
+        y = self._trajectory.get_y_positions()
+        z = self._trajectory.get_z_positions()
+
+        ax = plt.figure().add_subplot(projection='3d')
+        ax.scatter(x, y, z)
+
+        ax.set_xlabel('X')
+        ax.set_ylabel('Y')
+        ax.set_zlabel('Z')
+
+        plt.grid()
+        plt.show()
+
+    def get_trajectory(self) -> Trajectory:
+        return self._trajectory
+
+    def get_fps(self):
+        return self._fps
+
+    @abstractmethod
+    def _calculate_end_frame(self) -> int:
+        raise NotImplementedError
+
+    @abstractmethod
+    def _calculate_coordinates(self, frame: int) -> Coordinates:
+        raise NotImplementedError
+
+
+class CircleTrajectory(TrajectoryCalculator, ABC):
     _initial_motion_parameters: CircularTrajectoryMotionInitialParameters
     _motion_parameters: CircularTrajectoryMotionParameters
-    _fps: FramesPerSecond
 
     def __init__(self, initial_motion_parameters: CircularTrajectoryMotionInitialParameters, fps: FramesPerSecond):
         self._initial_motion_parameters = initial_motion_parameters
@@ -57,8 +66,21 @@ class CircleTrajectory(BasicBlenderTrajectoryCalculator, ABC):
     def get_motion_parameters(self) -> CircularTrajectoryMotionParameters:
         return self._motion_parameters
 
-    def get_fps(self):
-        return self._fps
 
-    def get_trajectory(self) -> Trajectory:
-        return self._trajectory
+class LinearTrajectory(TrajectoryCalculator, ABC):
+    _initial_motion_parameters: LinearTrajectoryMotionInitialParameters
+    _motion_parameters: LinearTrajectoryMotionParameters
+
+    def __init__(self,
+                 initial_motion_parameters: LinearTrajectoryMotionInitialParameters,
+                 fps: FramesPerSecond):
+        self._initial_motion_parameters = initial_motion_parameters
+        self._motion_parameters = LinearTrajectoryMotionParameters(self._initial_motion_parameters)
+        self._fps = fps
+        self._trajectory = Trajectory()
+
+    def get_initial_motion_parameters(self) -> LinearTrajectoryMotionInitialParameters:
+        return self._initial_motion_parameters
+
+    def get_motion_parameters(self) -> LinearTrajectoryMotionParameters:
+        return self._motion_parameters
