+++ blender3d_intergration/main.py
@@ -1,10 +1,13 @@
 import os
 
-from blender3d_intergration.object_trajectory_calculation.circle_trajectory import SimpleCircleTrajectory
-from blender3d_intergration.object_trajectory_calculation.models import Coordinates, \
-    CircularTrajectoryMotionInitialParameters, FramesPerSecond
+import numpy as np
 
-trajectories_path = './object_trajectory_calculation/calculated_trajectories/new'
+from blender3d_intergration.trajectories_api.circle_trajectory import SimpleCircleTrajectory
+from blender3d_intergration.trajectories_api.linear_trajectory import SimpleLinearTrajectory
+from blender3d_intergration.trajectories_api.models import Coordinates, \
+    CircularTrajectoryMotionInitialParameters, FramesPerSecond, LinearTrajectoryMotionInitialParameters
+
+trajectories_path = 'trajectories_api/calculated_trajectories/new'
 script_dir = os.path.dirname(__file__)
 full_path = os.path.join(script_dir, trajectories_path)
 
@@ -20,3 +23,16 @@ initial_params = CircularTrajectoryMotionInitialParameters(
 circular_trajectory = SimpleCircleTrajectory(initial_params, FramesPerSecond.FPS_30)
 circular_trajectory.calculate_trajectory()
 circular_trajectory.plot()
+
+linear_initial_params = LinearTrajectoryMotionInitialParameters(
+    start_point=Coordinates(x=10, y=0, z=30, frame=0, time_s=0),
+    end_point=Coordinates(x=100, y=0, z=30, frame=0, time_s=0),
+    x_velocity_m_s=18*np.sqrt(2),
+    y_velocity_m_s=0,
+    z_velocity_m_s=0
+)
+
+linear_trajectory = SimpleLinearTrajectory(linear_initial_params, FramesPerSecond.FPS_30)
+linear_trajectory.calculate_trajectory()
+linear_trajectory.plot()
+
